Relation fixes merge (#76)
* Init setup of patientplans

* Measurements change

* Remove TeamID column from HealthcareWorkers table and update related migrations; remove Measurements table from database schema.
Changes to relations for patient to other tables and workers. dataInit changes to fit relation changes , few tests fixed to fit new data and changes not all done

* data cast to decimal

* Update BodyWeight model to use float for Weight property
Update MeasurementsRepository to handle float values for Weight
Update HealthcareWorkerRepository to remove healthcare worker and related entities
Update BodyWeightRepository to fix calculation of VitalStatus

* Update BodyWeightDto and BodyWeightInputModel to use float for Weight property

* patientplan changes to measurement type of each to use frequency

* Bloodpressure PATCH Fix | Patient plan added team

+++ HomeVital.Repositories/Implementations/BodyWeightRepository.cs
@@ -52,6 +52,9 @@ namespace HomeVital.Repositories.Implementations{
             bodyWeight.PatientID = patientId;
             bodyWeight.Date = DateTime.UtcNow;
 
+            
+            
+
             _dbContext.BodyWeights.Add(bodyWeight);
             await _dbContext.SaveChangesAsync();
 
@@ -120,18 +123,18 @@ namespace HomeVital.Repositories.Implementations{
         /// <param name="currentWeight">The current body weight</param>
         /// <param name="bodyWeightRange">The body weight range</param>
         /// <returns>The status of the body weight</returns>
-        private string CheckBodyWeightRange(int id, double currentWeight, BodyWeightRange bodyWeightRange)
+        private string CheckBodyWeightRange(int id, float currentWeight, BodyWeightRange bodyWeightRange)
         {
             // Get the last 30 body weights of the patient
             var bodyWeights = _dbContext.BodyWeights
-            .Where(b => b.PatientID == id)
-            .OrderByDescending(b => b.Date)
-            .Take(30)
-            .ToList();
+                .Where(b => b.PatientID == id)
+                .OrderByDescending(b => b.Date)
+                .Take(30)
+                .ToList();
 
             if (bodyWeights.Count == 0)
             {
-            return VitalStatus.Normal.ToString();
+                return VitalStatus.Normal.ToString();
             }
 
             // // Calculate the average weight
